pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ajit Singh <ajit.singh2905@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
django = "^2.1"
pillow = "^5.3"
django-imagekit = "^4.0"
djangorestframework = "^3.9"
markdown = "^3.0"
django-cors-headers = "^2.4"
djangorestframework_simplejwt = "^3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
